endday10

+++ posn2/working/dp/dp0_2.cpp
@@ -4,29 +4,29 @@ using namespace std;
 
 int main(){
 
-    int m,n;
+    long long m,n;
     cin >> m >> n;
 
-    vector <vector <int>> v(m+1, vector<int>(n+1, 0));
+    vector <vector <long long>> v(m+1, vector<long long>(n+1, 0));
 
-    for (int i = 1; i <= m; i++){
-        for (int j = 1; j <= n; j++){
+    for (long long i = 1; i <= m; i++){
+        for (long long j = 1; j <= n; j++){
             cin >> v[i][j];
         }
     }
 
 
     // loop จะตาม row
-    for (int i = 1; i <= m; i++){
-        for (int j = 1; j <= n; j++){
+    for (long long i = 1; i <= m; i++){
+        for (long long j = 1; j <= n; j++){
             v[i][j] += max(v[i-1][j-1], v[i-1][j]);
         }
     }
 
     cout << v[m][n] << '\n';
 
-    for (int i = 1; i <= m; i++){
-        for (int j = 1; j <= n; j++){
+    for (long long i = 1; i <= m; i++){
+        for (long long j = 1; j <= n; j++){
             cout << v[i][j] << " ";
         }
         cout << '\n';
